refactor: 포인트 입력 로직 커스텀 훅 분리
