+++ src/components/PriceInput.tsx
@@ -1,7 +1,5 @@
-import { useState } from 'react';
-import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
-import cartOrderPriceState from '../recoil/selectors/cartOrderPriceState';
+import usePointInput from '../hooks/usePointInput';
 
 const InputContainer = styled.div`
   display: flex;
@@ -48,24 +46,7 @@ interface PriceInputProps {
 
 const PriceInput = (props: PriceInputProps) => {
   const { useablePoint, handleInputField } = props;
-  const [inputValue, setInputValue] = useState('');
-  const prices = useRecoilValue(cartOrderPriceState);
-
-  const checkPointInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (+e.target.value % 10 !== 0) {
-      e.target.value = `${Math.floor(+e.target.value / 10).toString()}0`;
-    }
-
-    if (+e.target.value > useablePoint) {
-      e.target.value = `${Math.floor(+useablePoint / 10).toString()}0`;
-    }
-
-    if (+e.target.value > prices.total) {
-      e.target.value = `${Math.floor(+prices.total / 10).toString()}0`;
-    }
-
-    setInputValue(e.target.value);
-  };
+  const { checkPointInput, inputValue } = usePointInput(useablePoint);
 
   const usePoint = () => {
     handleInputField(+inputValue);
